backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
PyMySQL==1.1.0
pyodbc==5.0.1
SQLAlchemy==2.0.23
Werkzeug==3.0.1
gunicorn==21.2.0
geopy==2.4.1
requests==2.31.0
